timer.performwithdelay
the timer.performwithdelay could be executed even if the level was
cancelled - hitting home - replace with a tagged timer on ‘win’ which
fires the win game event after 4 seconds, but is cancelled if the home
button is pressed. The timer event isn’t cleared because the scene
manager object exists throughout, so it can be fired next time around.
Gotcha :)